Updated code

+++ src/main/java/com/codearchitects/todoapp/Repositories/UserRepository.java
@@ -7,11 +7,15 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     @Query("SELECT u FROM User u WHERE u.userName=:un")
     User findByUserName(@Param("un") String userName);
 
     Boolean existsByEmail(String email);
 }
+
+
+
+
